fix compile error

+++ source/jucePlugin/ui3/VirusEditor.cpp
@@ -297,7 +297,7 @@ namespace genericVirusUI
 		const auto& props = _component->getProperties();
 		const int v = props["parameter"];
 
-		const int part = props.contains("part") ? props["part"] : getController().getCurrentPart();
+		const int part = props.contains("part") ? static_cast<int>(props["part"]) : static_cast<int>(getController().getCurrentPart());
 
 		const auto* p = getController().getParameter(static_cast<Virus::ParameterType>(v), part);
